Class fields lesson: add a comma for extra clarity
While a friend of mine was going through the documentation, since they are not a native English speaker, they reached out to me and asked about that specific part. I realized the lack of a comma made it appear less clear than it should.

This is a simple fix that will make the sentence a bit clearer, especially for non-native readers.